# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FbTokenStatus = (typeof FbTokenStatus)[keyof typeof FbTokenStatus];

export const FbTokenStatus = {
  active: "active",
  expired: "expired",
  invalid: "invalid",
} as const;

export interface FbToken {
  id: number;
  fbId: string;
  token: string;
  cookie?: string | null;
  status: FbTokenStatus;
  note?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateFbTokenBodyStatus =
  (typeof CreateFbTokenBodyStatus)[keyof typeof CreateFbTokenBodyStatus];

export const CreateFbTokenBodyStatus = {
  active: "active",
  expired: "expired",
  invalid: "invalid",
} as const;

export interface CreateFbTokenBody {
  fbId: string;
  token: string;
  cookie?: string | null;
  status: CreateFbTokenBodyStatus;
  note?: string | null;
}

export type UpdateFbTokenBodyStatus =
  (typeof UpdateFbTokenBodyStatus)[keyof typeof UpdateFbTokenBodyStatus];

export const UpdateFbTokenBodyStatus = {
  active: "active",
  expired: "expired",
  invalid: "invalid",
} as const;

export interface UpdateFbTokenBody {
  fbId?: string;
  token?: string;
  cookie?: string | null;
  status?: UpdateFbTokenBodyStatus;
  note?: string | null;
}

export interface TokenStats {
  total: number;
  active: number;
  expired: number;
  invalid: number;
}

export type ListTokensParams = {
  status?: ListTokensStatus;
  search?: string;
};

export type ListTokensStatus =
  (typeof ListTokensStatus)[keyof typeof ListTokensStatus];

export const ListTokensStatus = {
  active: "active",
  expired: "expired",
  invalid: "invalid",
} as const;