# lib/api-zod/src/generated/types/createFbTokenBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateFbTokenBodyStatus } from "./createFbTokenBodyStatus";

export interface CreateFbTokenBody {
  fbId: string;
  token: string;
  cookie?: string | null;
  status: CreateFbTokenBodyStatus;
  note?: string | null;
}