# lib/api-zod/src/generated/types/fbToken.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FbTokenStatus } from "./fbTokenStatus";

export interface FbToken {
  id: number;
  fbId: string;
  token: string;
  cookie?: string | null;
  status: FbTokenStatus;
  note?: string | null;
  createdAt: Date;
  updatedAt: Date;
}