# lib/api-zod/src/generated/types/updateFbTokenBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateFbTokenBodyStatus } from "./updateFbTokenBodyStatus";

export interface UpdateFbTokenBody {
  fbId?: string;
  token?: string;
  cookie?: string | null;
  status?: UpdateFbTokenBodyStatus;
  note?: string | null;
}